fixed lint

+++ posting/mvcc_test.go
@@ -113,10 +113,10 @@ func BenchmarkTestCache(b *testing.B) {
 
 	attr := x.GalaxyAttr("cache")
 	keys := make([][]byte, 0)
-	N := uint64(10000)
+	N := 10000
 	txn := Oracle().RegisterStartTs(1)
 
-	for i := uint64(1); i < N; i++ {
+	for i := uint64(1); i < uint64(N); i++ {
 		key := x.DataKey(attr, i)
 		keys = append(keys, key)
 		edge := &pb.DirectedEdge{
@@ -143,7 +143,7 @@ func BenchmarkTestCache(b *testing.B) {
 
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			key := keys[rand.Intn(int(N-1))]
+			key := keys[rand.Intn(N-1)]
 			_, err = getNew(key, pstore, math.MaxUint64)
 			if err != nil {
 				panic(err)
